test(fargate): split bom into its own test file
